% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansTestPr.R
\name{meansTestP}
\alias{meansTestP}
\title{Significance test for difference between 2 mean values}
\usage{
meansTestP(av1, sd1, nd1, av2, sd2, nd2, rg1, rg2)
}
\arguments{
\item{av1}{mean value of group 1}

\item{sd1}{standard deviation of group 1}

\item{nd1}{number of group 1}

\item{av2}{mean value of group 2}

\item{sd2}{standard deviation of group 2}

\item{nd2}{number of group 2}

\item{rg1}{plot axis min}

\item{rg2}{plot axis max}
}
\value{
Deviation Test
}
\description{
\code{meansTestP} statistical test for mean values
}
\examples{
# meansTestP(50, 10, 20, 60, 15, 25, 0, 100)
}
